refactor: '경로 index.js 생략' 커밋의 revert
